refactor: extract buildMenubarPayloadForRange for reuse by MCP
